requirements.txt
blinker==1.9.0
Bootstrap-Flask==2.4.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-MySQLdb==2.0.0
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysqlclient==2.2.7
platformdirs==4.3.6
PyJWT==2.10.1
pymongo==4.11.1
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
virtualenv==20.26.6
Werkzeug==3.1.3
WTForms==3.2.1
